js: Wait until ExternalInterface is ready for use.
When flashcanvas.js is loaded, flashcanvas.swf is also loaded by a
synchronous Ajax request and cached in the browser. This ensures that
postCommands() and resize() are available from the start.

In some situations, however, JavaScript fails to call these ActionSctipt
functions, probably because ExternalInterface is not ready for use. Just
to be safe, we should check that the swf is ready before calling
ActionScript functions.